Seperate sd module into smaller files
