+++ src/sd/sd_event_group_entry_type.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum SdEventGroupEntryType {
+    Subscribe = 0x06,
+    SubscribeAck = 0x07,
+}
