+++ src/sd/sd_service_entry_type.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum SdServiceEntryType {
+    FindService = 0x00,
+    OfferService = 0x01,
+}
